tasks feature updated. toast package installed

+++ components/tasksPage/DeleteTaskBtn.jsx
@@ -1,14 +1,23 @@
+// context 
+import { useGlobalContext } from "@/app/context";
 // lib - actions
 import { deleteTaskAction } from "@/lib/actions/taskActions"
 // icon
 import { MdDelete } from "react-icons/md";
+// toast
+import toast from "react-hot-toast";
 
-const DeleteTaskBtn = ({ userID, taskID, getTasksList }) => {
+const DeleteTaskBtn = ({ userID, taskID }) => {
+    const { tasks, setTasks } = useGlobalContext()
+
     const handleDeleteTask = async () => {
-        const response = await deleteTaskAction(userID, taskID)
-        if (response) {
-            getTasksList()
-            console.log('Selected task has been removed')
+        if (window.confirm('Are you sure you want to delete task')) {
+            const response = await deleteTaskAction(userID, taskID)
+            if (response) {
+                const updatedTasksList = tasks.filter(task => task.docID !== taskID)
+                setTasks(updatedTasksList)
+                toast.success('Selected task has been removed')
+            }
         }
     }
